Require `json` or `xml` in provider cores
Since `fog-json` was extracted from `fog` it means if you rely on:

    require "fog/provider/compute"

`Fog::JSON` was not defined although it was supposed to be a supported
use case.

This adds the requires for JSON or XML based APIs based on a quick scan
of each provider to declare the required part.

AWS seems to be relying on both. Which is nice.

Providers relying on native code dependencies