Fix DNS record updates for fog-core 1.25.0 and later
When PUT is called for a Dynect CNAME record without specifying
a record_id, the response contains an array of records.  The
original code for supporting CNAME updates assumed that a
single record would be returned.  In fog-core prior to 1.25.0,
merge_attributes was quietly mishandling this data without
raising an exception.  Starting with version 1.25.0, the use
of 'each_pair' to iterate the hash caused an exception to be
raised when an array was passed instead.

This change fixes PUT calls by first retrieving the existing
record and performing the update on the correct record_id.